requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via
    #   aiohttp-retry
    #   twilio
aiohttp-retry==2.9.1
    # via twilio
aiosignal==1.4.0
    # via aiohttp
amqp==5.3.1
    # via kombu
anyio==4.10.0
    # via httpx
asgiref==3.9.1
    # via django
async-timeout==5.0.1
    # via aiohttp
attrs==25.3.0
    # via aiohttp
billiard==4.2.1
    # via celery
celery==5.5.3
    # via -r requirements.in
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via rabbitmq
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via click
django==5.2.5
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.16.1
    # via -r requirements.in
exceptiongroup==1.3.0
    # via anyio
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via python-telegram-bot
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
kombu==5.5.4
    # via celery
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
packaging==25.0
    # via kombu
pillow==11.3.0
    # via -r requirements.in
prompt-toolkit==3.0.51
    # via click-repl
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
psycopg2==2.9.10
    # via -r requirements.in
pycparser==2.22
    # via cffi
pyjwt==2.10.1
    # via twilio
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.1.1
    # via -r requirements.in
python-telegram-bot==22.3
    # via -r requirements.in
rabbitmq==0.2.0
    # via -r requirements.in
requests==2.32.4
    # via twilio
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlparse==0.5.3
    # via django
twilio==9.7.0
    # via -r requirements.in
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   anyio
    #   asgiref
    #   exceptiongroup
    #   multidict
tzdata==2025.2
    # via
    #   django
    #   kombu
urllib3==2.5.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
yarl==1.20.1
    # via aiohttp
